remove the internal ips from default config
