+++ polaris-common/polaris-config-default/src/main/resources/META-INF/services/com.tencent.polaris.api.config.ConfigProvider
@@ -1,0 +1,1 @@
+com.tencent.polaris.config.internal.DefaultConfigProvider
