--- polaris-common/polaris-config-default/src/main/resources/META-INF/services/com.tencent.polaris.api.config.DefaultConfigProvider
@@ -1,1 +1,0 @@
-com.tencent.polaris.config.internal.DefaultInternalConfigProvider
